Prevent race condition in `get_with` methods to avoid evaluating `init` closure/future multiple times
In the current implementation, there is a small chance that the following methods to
evaluate the `init` closure/function more than once in concurrent calls:

- `get_with`
- `get_with_if`
- `optionally_get_with`
- `try_get_with`

It will happen when a call to one of these methods is made at the time when another
call to the same method is about to complete.

New test cases in `tests/entry_api_sync.rs` will reproduce the issue.

This commit contains the fix by doing a double check if the value still does not
exist after the lock is acquired and also extending the lock as needed.