Upgrade thiserror to 2.0
Thiserror has had a 2.0 release that does include breaking changes but
does not appear to impact moka's use of thiserror.

By upgrading, we can stay closer to the latest version of thiserror and
avoid needing to compile multiple versions of thiserror in consumers who
are themselves using a newer version of thiserror.